Rename processor_registry.py to processor_factory.py
