+++ explainaboard/processors/__init__.py
@@ -27,7 +27,7 @@ from explainaboard.processors import (
     text_pair_classification,
     word_segmentation,
 )
-from explainaboard.processors.processor_registry import get_processor
+from explainaboard.processors.processor_factory import get_processor
 
 __all__ = [
     'aspect_based_sentiment_classification',
