+++ explainaboard/processors/grammatical_error_correction_test.py
@@ -8,7 +8,7 @@ from explainaboard import TaskType
 from explainaboard.processors.grammatical_error_correction import (
     GrammaticalErrorCorrectionProcessor,
 )
-from explainaboard.processors.processor_registry import get_processor
+from explainaboard.processors.processor_factory import get_processor
 from explainaboard.serialization.serializers import PrimitiveSerializer
