+++ explainaboard/processors/qa_multiple_choice_test.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import unittest
 
 from explainaboard import TaskType
-from explainaboard.processors.processor_registry import get_processor
+from explainaboard.processors.processor_factory import get_processor
 from explainaboard.processors.qa_multiple_choice import QAMultipleChoiceProcessor
 from explainaboard.serialization.serializers import PrimitiveSerializer
